Create Glue crawler created.py
This code initializes a session with AWS using specified credentials, creates a Glue client, and defines a Glue crawler named ev-data-crawler for scanning data stored in the S3 bucket electricvehiclesrawdata. The crawler is associated with the Glue database electricvehicles_database, and its target is the raw-data folder in the S3 bucket. Finally, the code creates and starts the Glue crawler to populate the Glue Data Catalog with metadata from the specified S3 path.