Create Upload data to S3.py
This code defines a function to upload a file to an S3 bucket using the boto3 library. The function takes the file name, S3 bucket name, and an optional S3 object name as inputs. It initializes an S3 client, attempts to upload the file, and handles potential errors such as file not found, missing credentials, or other exceptions. If successful, it confirms the upload; otherwise, it prints the appropriate error message. Finally, the code uploads the specified file Electric_Vehicle_Population_Data.csv to the electricvehiclesrawdata S3 bucket.